# api/swagger/api/todo-service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ToDo REST API",
    "version": "1.0",
    "contact": {
      "name": "go-grpc-http-rest-microservice",
      "url": "https://github.com/metehanakbaba/go-grpc-http-rest-microservice-tutorial"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/todo/all": {
      "get": {
        "operationId": "ReadAll",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiReadAllResponse"
            }
          },
          "404": {
            "description": "Endpoint bulunamadı.",
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        },
        "tags": [
          "ToDoService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "apiReadAllResponse": {
      "type": "object",
      "properties": {
        "toDos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiToDo"
          }
        }
      }
    },
...
    "apiToDo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64",
          "title": "Benzersiz integer ID (A_I)"
        },
        "title": {
          "type": "string",
          "title": "Task Basligi ToDo Task icin"
        },
        "description": {
          "type": "string",
          "title": "Detayli Aciklama ToDo Task icin"
        },
        "reminder": {
          "type": "string",
          "format": "date-time",
          "title": "Olusturlacak ToDo Task icin hatirlatma (Tarih/Saat) timestamp formatında"
        }
      },
      "title": "ToDo Task"
    },
...
  }
}